+++ src/corpus/sources/SourcesPanel.vue
@@ -80,7 +80,7 @@ async function fileHandler(files: File[], onProgress: ProgressHandler) {
       <table v-if="sources.length" class="w-full mt-4 striped">
         <thead>
           <tr>
-            <th class="w-full">{{ $t("fileName") }}</th>
+            <th class="w-full">{{ $t("filename") }}</th>
             <th class="text-right">{{ $t("fileSize") }}</th>
             <th v-if="canWrite('corpora', corpusId)" class="sr-only">
               {{ $t("file.operations") }}