+++ CHANGELOG.md
@@ -13,6 +13,10 @@ As this project is a user-facing application, the places in the semantic version
 
 ## [Unreleased](https://github.com/spraakbanken/mink-frontend/compare/v1.15.1...HEAD)
 
+### Added
+
+- CodeMirror editor with syntax highlighting and inline schema validation for custom config YAML
+
 ## [1.15.1](https://github.com/spraakbanken/mink-frontend/releases/tag/v1.15.1) (2026-02-04)
 
 ### Fixed